move the viewModels to another package
