+++ app/src/main/java/com/example/helphero/ui/signin/signInFragment.kt
@@ -12,10 +12,11 @@ import com.example.helphero.R
 import com.example.helphero.databases.users.UserDatabase
 import com.example.helphero.databinding.FragmentSignInBinding
 import com.example.helphero.repositories.UserRepository
-import com.example.helphero.viewmodels.SignInViewModelFactory
+import com.example.helphero.ui.viewmodels.SignInViewModelFactory
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.example.helphero.MainActivity
+import com.example.helphero.ui.viewmodels.SignInViewModel
 
 class SignInFragment : Fragment() {
     private var _binding: FragmentSignInBinding? = null
