+++ app/src/main/java/com/example/helphero/ui/viewmodels/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.helphero.ui.signup
+package com.example.helphero.ui.viewmodels
 
 import android.util.Log
 import androidx.lifecycle.LiveData
